rework component structure
