+++ components/client/ClientLocationPicker.js
@@ -8,23 +8,19 @@ import {
   TouchableOpacity,
   TextInput,
 } from "react-native";
+import {
+  PassSetLocation,
+  PassLocationState,
+} from "../../screens/client/NewProjectScreenOne";
 import { useNavigation } from "@react-navigation/native";
-import { APP_STRINGS } from "../constants/index";
+import { APP_STRINGS } from "../../constants/index";
 
-//REFACTORED with APP_STRINGS and TURNARY VIA FRANKS SPECIFICATIONS
-
-const {
-  workExperienceSet,
-  setWorkExperience,
-  choose,
-  briefSummary,
-  slide
-} = APP_STRINGS;
-
-const BioPicker = (props) => {
-  const { personalBio, setPersonalBio } = props;
+const ClientLocationPicker = () => {
+  const navigation = useNavigation();
   const [isModalVisible, setIsModalVisible] = useState(false);
 
+  const setLocation = useContext(PassSetLocation);
+  const locationState = useContext(PassLocationState);
 
   const openModal = () => {
     setIsModalVisible(true);
@@ -34,51 +30,56 @@ const BioPicker = (props) => {
     setIsModalVisible(false);
   };
 
-  const renderPersonalBioButton = (buttonText = "") => {
-    const title = openModal;
+  const renderClientLocationButton = (buttonText) => {
     return (
-      <TouchableOpacity style={styles.button} onPress={openModal} title={title}>
+      <TouchableOpacity
+        style={styles.button}
+        onPress={openModal}
+        title={"Open modal"}
+      >
         <Text style={styles.buttonText}>{buttonText}</Text>
       </TouchableOpacity>
     );
   };
 
-  const renderPersonalBio = (hasPersonalBio = false) => {
-    return hasPersonalBio
-      ? renderPersonalBioButton(workExperienceSet)
-      : renderPersonalBioButton(setWorkExperience);
+  const renderClientLocation = (hasClientLocation) => {
+    return hasClientLocation
+      ? renderClientLocationButton(locationState)
+      : renderClientLocationButton("Please Set Location");
   };
 
-  const renderTextInput = (bio, setBio) => {
+  const renderTextInput = () => {
     return (
       <View>
-        <TextInput style={styles.input} onChangeText={setBio} value={bio} />
+        <TextInput
+          style={styles.input}
+          maxLength={26}
+          onChangeText={setLocation}
+          value={locationState}
+        />
       </View>
     );
   };
 
   return (
-    <View>
+    <View style={styles.container}>
       <Modal
         transparent={true}
         visible={isModalVisible}
-        animationType={slide}
+        animationType={APP_STRINGS.slide}
         onRequestClose={closeModal}
       >
         <View style={styles.modalContainer}>
           <View style={styles.innerContainer}>
-            <Text style={styles.modalText}>{briefSummary}</Text>
+            <Text style={styles.modalText}>{APP_STRINGS.whereIsTheLocation}</Text>
           </View>
-
-          {renderTextInput(personalBio, setPersonalBio)}
-
+          {renderTextInput()}
           <View styles={styles.cancelWrapper}>
-            <Button onPress={closeModal} title={choose} />
+            <Button onPress={closeModal} title={"Choose"} />
           </View>
         </View>
       </Modal>
-
-      {renderPersonalBio(personalBio)}
+      {renderClientLocation(locationState)}
     </View>
   );
 };
@@ -120,6 +121,7 @@ const styles = StyleSheet.create({
   },
   buttonText: {
     color: "white",
+    fontSize: 20,
   },
   input: {
     marginTop: 20,
@@ -130,4 +132,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default BioPicker;
+export default ClientLocationPicker;
