+++ components/pilot/DroneExperiencePicker.js
@@ -11,8 +11,8 @@ import {
 import {
   PassSetYearsOfExperience,
   PassYearsOfExperienceState,
-} from "../screens/pilot/PilotProfileSetupPageOneScreen";
-import { APP_STRINGS } from "../constants/index";
+} from "../../screens/pilot/PilotProfileSetupPageOneScreen";
+import { APP_STRINGS } from "../../constants/index";
 
 //REFACTORED with APP_STRINGS and TURNARY VIA FRANKS SPECIFICATIONS
