+++ components/shared/MainHeader.js
@@ -12,15 +12,15 @@ import {
   Alert,
 } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
-import hovrtekLogo from "../assets/hovrtek_logo.png";
+import hovrtekLogo from "../../assets/hovrtek_logo.png";
 import { useNavigation, DrawerActions } from "@react-navigation/native";
 import Notification from "./Notification";
-import { getMessages } from "../actions/messages.js";
+import { getMessages } from "../../actions/messages.js";
 import { connect } from "react-redux";
 import _ from "lodash";
 import * as firebase from "firebase";
-import { NotificationContext } from "../context";
-import { HeaderContext } from "../context";
+import { NotificationContext } from "../../context";
+import { HeaderContext } from "../../context";
 
 const MainHeader = (props, { getMessages }) => {
   const navigation = useNavigation();
