+++ components/shared/NestedHeader.js
@@ -6,7 +6,7 @@ import {
   Platform,
 } from "react-native";
 import { Ionicons } from "@expo/vector-icons";
-import hovrtekLogo from "../assets/hovrtek_logo.png";
+import hovrtekLogo from "../../assets/hovrtek_logo.png";
 
 const NestedHeader = () => {
   return (
