+++ navigation/MessagingNavigation.js
@@ -2,8 +2,8 @@ import React from "react";
 import { createStackNavigator } from "@react-navigation/stack";
 import MessagingScreen from "../screens/messaging/MessagingScreen";
 import ChatScreen from "../screens/messaging/ChatScreen";
-import NestedHeader from '../components/NestedHeader';
-import MainHeader from '../components/MainHeader';
+import NestedHeader from '../components/shared/NestedHeader';
+import MainHeader from '../components/shared/MainHeader';
 import { Ionicons } from "@expo/vector-icons";
 
 const MessagingNavigatorStack = createStackNavigator();
